% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddSlices_Rcpp}
\alias{AddSlices_Rcpp}
\title{Add slices and return vector of the sums of slices}
\usage{
AddSlices_Rcpp(slicematrix)
}
\arguments{
\item{slicematrix}{A matrix of slices.}
}
\value{
Returns the vector of the sums of the slices
}
\description{
Add slices and return vector of the sums of slices
}
