% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddWeightedSlices_Rcpp}
\alias{AddWeightedSlices_Rcpp}
\title{Add slices and return vector of the weighted sums of slices}
\usage{
AddWeightedSlices_Rcpp(slicematrix, weightvector)
}
\arguments{
\item{slicematrix}{A matrix of slices.}

\item{weightvector}{A vector of weights.}
}
\value{
Returns the vector of the weighted sums of the slices
}
\description{
Add slices and return vector of the weighted sums of slices
}
