% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altitude.R
\name{ambit_get_altitude}
\alias{ambit_get_altitude}
\title{Extract altitude data from an openambit log file}
\usage{
ambit_get_altitude(log_file)
}
\arguments{
\item{log_file}{The path to an openambit log file}
}
\value{
A data.frame with two columns \code{altitude} and
   \code{distance} containing the logs for altitude by distance within
   the track.
}
\description{
Extract altitude data from an openambit log file
}
