% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gps.R
\name{ambit_get_gps}
\alias{ambit_get_gps}
\title{Extract GPS data from an openambit log file}
\usage{
ambit_get_gps(log_file)
}
\arguments{
\item{log_file}{The path to an openambit log file}
}
\value{
A data.frame with two columns \code{lon} and
   \code{lat} containing the GPS logs for longitude and latitude,
   respectively.
}
\description{
Extract GPS data from an openambit log file
}
