% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions.R
\name{my_mae}
\alias{my_mae}
\title{my_mse}
\usage{
my_mae(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
Mean absolute error between the two vectors x and y
}
\description{
Returns the mean absolute error between two vectors
}
\examples{
#Simulate two vectors of i.i.d.~standard normal data
set.seed(456)
x <- rnorm(100)
y <- rnorm(100)
#Compute the mean absolute error between both vectors
my_mae(x,y)
}
