% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrawlFunctions-trawl.R
\name{trawl_Exp}
\alias{trawl_Exp}
\title{Evaluates the exponential trawl function}
\usage{
trawl_Exp(x, lambda)
}
\arguments{
\item{x}{the argument at which the exponential trawl function will be
evaluated}

\item{lambda}{the parameter \eqn{\lambda} in the exponential trawl}
}
\value{
The exponential trawl function evaluated at x
}
\description{
Evaluates the exponential trawl function
}
\details{
The trawl function is parametrised by parameter \eqn{\lambda > 0} as
 follows: \deqn{g(x) = e^{\lambda x},  \mbox{ for }  x \le 0.}
}
\examples{
trawl_Exp(-1,0.5)
}
